chore: renamed: lval->symtable
