+++ symtable.h
